Change folder naming and distribution
Changed folder naming and distribution to reflect the use of mvp
architecure and therefore simplify the first contact of non-android
developers with this repository (users should be familiarized with MVP
pattern).
