--- app/src/main/java/br/org/cesar/knot_setup_app/fragment/configureDevice/ConfigureDeviceContract.java
@@ -1,16 +1,0 @@
-package br.org.cesar.knot_setup_app.fragment.configureDevice;
-
-public interface ConfigureDeviceContract {
-
-    interface ViewModel {
-        void onConnected();
-        void onDisconnected();
-        void onWriteSucceeded(int write_count);
-        void onWriteFailed(int write_count);
-        void onErrorHandler(Throwable throwable);
-    }
-
-    interface Presenter {
-        void onFocus();
-    }
-}
