--- app/src/main/java/br/org/cesar/knot_setup_app/fragment/configureGatewayWifi/ConfigureGatewayWifiContract.java
@@ -1,15 +1,0 @@
-package br.org.cesar.knot_setup_app.fragment.configureGatewayWifi;
-
-public interface ConfigureGatewayWifiContract {
-
-    interface ViewModel {
-        void setWifiSSID(String wifiName);
-        void onDisconnected();
-        void onWriteFailed();
-    }
-
-    interface Presenter {
-        void onFocus();
-        void onSetWifiClicked(String ssid, String pwd);
-    }
-}
