--- app/src/main/java/br/org/cesar/knot_setup_app/fragment/gatewayConnected/GatewayConnectedContract.java
@@ -1,19 +1,0 @@
-package br.org.cesar.knot_setup_app.fragment.gatewayConnected;
-
-import android.net.nsd.NsdServiceInfo;
-
-import java.util.ArrayList;
-
-public interface GatewayConnectedContract {
-
-    interface ViewModel {
-        void onGatewaysFound(ArrayList<NsdServiceInfo> deviceList);
-        void setSearchingFeedback(int visibility);
-    }
-
-    interface Presenter {
-        void onFocusLost();
-        void onGatewayClicked(NsdServiceInfo serviceInfo);
-        void onFocus();
-    }
-}
