+++ app/src/main/java/br/org/cesar/knot_setup_app/view/PageAdapter.java
@@ -1,0 +1,37 @@
+package br.org.cesar.knot_setup_app.view;
+
+import android.support.v4.app.Fragment;
+import android.support.v4.app.FragmentManager;
+import android.support.v4.app.FragmentPagerAdapter;
+
+import java.util.ArrayList;
+
+public class PageAdapter extends FragmentPagerAdapter {
+
+    private ArrayList<Fragment> fragmentList = new ArrayList<>();
+    private ArrayList<String> fragmentTitleList = new ArrayList<>();
+
+    public PageAdapter(FragmentManager fm) {
+        super(fm);
+    }
+
+    public void addList(Fragment fragment, String title){
+        fragmentList.add(fragment);
+        fragmentTitleList.add(title);
+    }
+
+    @Override
+    public CharSequence getPageTitle(int position) {
+        return fragmentTitleList.get(position);
+    }
+
+    @Override
+    public Fragment getItem(int position) {
+        return fragmentList.get(position);
+    }
+
+    @Override
+    public int getCount() {
+        return fragmentList.size();
+    }
+}
