+++ app/src/main/java/br/org/cesar/knot_setup_app/view/login/LoginContract.java
@@ -1,0 +1,15 @@
+package br.org.cesar.knot_setup_app.view.login;
+
+public interface LoginContract {
+
+    interface ViewModel {
+        void fillEmail(String email);
+        void onInvalidCredentials();
+        void onLogin();
+    }
+
+    interface Presenter {
+        void onFocus();
+        void onSendClicked(String login, String pwd);
+    }
+}
