+++ app/src/main/java/br/org/cesar/knot_setup_app/view/splash/SplashContract.java
@@ -1,0 +1,15 @@
+package br.org.cesar.knot_setup_app.view.splash;
+
+public interface SplashContract {
+
+    interface Presenter {
+        void chooseActivity();
+    }
+
+    interface ViewModel{
+        void doLogin();
+        void doListGateways();
+        void apiNotConfigured();
+        void callbackOnConnectionError();
+    }
+}
