+++ app/src/main/java/br/org/cesar/knot_setup_app/view/thingRegistered/ThingContract.java
@@ -1,0 +1,16 @@
+package br.org.cesar.knot_setup_app.view.thingRegistered;
+
+import java.util.List;
+
+import br.org.cesar.knot_setup_app.model.Gateway;
+
+public interface ThingContract {
+
+    interface ViewModel {
+        void onThingsFound(List<Gateway> deviceList);
+    }
+
+    interface Presenter {
+        void onFocus();
+    }
+}
