+++ app/src/main/res/layout/fragment_configure_gateway_wifi.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorWhite"
-    tools:context=".activity.ConfigureGatewayWifiActivity">
+    tools:context=".view.ConfigureGatewayWifiActivity">
 
 
     <TextView
