+++ app/src/main/res/layout/fragment_configure_thing.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:background="@color/colorWhite"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".activity.ConfigureDeviceActivity">
+    tools:context=".view.ConfigureDeviceActivity">
 
     <include
         android:id="@+id/ot_frame"
